Introduce a provider package and move Azure code there
