+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/command/PullRequestOpenCommand.kt
@@ -1,8 +1,8 @@
 package io.rabobank.ret.git.plugin.command
 
 import io.quarkus.logging.Log
-import io.rabobank.ret.git.plugin.azure.AzureDevopsClient
-import io.rabobank.ret.git.plugin.azure.AzureDevopsUrlFactory
+import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsClient
+import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsUrlFactory
 import io.rabobank.ret.git.plugin.output.OutputHandler
 import io.rabobank.ret.picocli.mixin.ContextAwareness
 import io.rabobank.ret.util.BrowserUtils
