+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/output/AlfredAutocompleteHandler.kt
@@ -2,13 +2,13 @@ package io.rabobank.ret.git.plugin.output
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.rabobank.ret.RetConsole
-import io.rabobank.ret.git.plugin.azure.Branch
-import io.rabobank.ret.git.plugin.azure.Pipeline
-import io.rabobank.ret.git.plugin.azure.PipelineRun
-import io.rabobank.ret.git.plugin.azure.PipelineRunResult
-import io.rabobank.ret.git.plugin.azure.PipelineRunState
-import io.rabobank.ret.git.plugin.azure.PullRequest
-import io.rabobank.ret.git.plugin.azure.Repository
+import io.rabobank.ret.git.plugin.provider.azure.Branch
+import io.rabobank.ret.git.plugin.provider.azure.Pipeline
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRun
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRunResult
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRunState
+import io.rabobank.ret.git.plugin.provider.azure.PullRequest
+import io.rabobank.ret.git.plugin.provider.azure.Repository
 
 class AlfredAutocompleteHandler(private val retConsole: RetConsole, private val objectMapper: ObjectMapper) :
     OutputHandler {
