+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/output/AlfredOutputDomain.kt
@@ -1,9 +1,9 @@
 package io.rabobank.ret.git.plugin.output
 
 import io.quarkus.runtime.annotations.RegisterForReflection
-import io.rabobank.ret.git.plugin.azure.Branch
-import io.rabobank.ret.git.plugin.azure.PullRequest
-import io.rabobank.ret.git.plugin.azure.Repository
+import io.rabobank.ret.git.plugin.provider.azure.Branch
+import io.rabobank.ret.git.plugin.provider.azure.PullRequest
+import io.rabobank.ret.git.plugin.provider.azure.Repository
 
 @RegisterForReflection
 data class Wrapper(val items: List<Item>)
