+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/output/CliAutocompleteHandler.kt
@@ -1,12 +1,12 @@
 package io.rabobank.ret.git.plugin.output
 
 import io.rabobank.ret.RetConsole
-import io.rabobank.ret.git.plugin.azure.Branch
-import io.rabobank.ret.git.plugin.azure.Pipeline
-import io.rabobank.ret.git.plugin.azure.PipelineRun
-import io.rabobank.ret.git.plugin.azure.PipelineRunState
-import io.rabobank.ret.git.plugin.azure.PullRequest
-import io.rabobank.ret.git.plugin.azure.Repository
+import io.rabobank.ret.git.plugin.provider.azure.Branch
+import io.rabobank.ret.git.plugin.provider.azure.Pipeline
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRun
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRunState
+import io.rabobank.ret.git.plugin.provider.azure.PullRequest
+import io.rabobank.ret.git.plugin.provider.azure.Repository
 
 class CliAutocompleteHandler(private val retConsole: RetConsole) : OutputHandler {
     override fun listPRs(list: List<PullRequest>) {
