+++ git-plugin/src/main/kotlin/io/rabobank/ret/git/plugin/provider/azure/AuthorizationHeaderInjector.kt
@@ -1,4 +1,4 @@
-package io.rabobank.ret.git.plugin.azure
+package io.rabobank.ret.git.plugin.provider.azure
 
 import io.rabobank.ret.git.plugin.config.PluginConfig
 import jakarta.inject.Singleton
