+++ git-plugin/src/test/kotlin/io/rabobank/ret/git/plugin/AutoCompleteCommandTest.kt
@@ -5,16 +5,16 @@ import io.rabobank.ret.IntelliSearch
 import io.rabobank.ret.RetContext
 import io.rabobank.ret.configuration.Configurable
 import io.rabobank.ret.configuration.RetConfig
-import io.rabobank.ret.git.plugin.azure.AzureDevopsClient
-import io.rabobank.ret.git.plugin.azure.AzureResponse
-import io.rabobank.ret.git.plugin.azure.Branch
-import io.rabobank.ret.git.plugin.azure.Pipeline
-import io.rabobank.ret.git.plugin.azure.PipelineRun
-import io.rabobank.ret.git.plugin.azure.PipelineRunResult
-import io.rabobank.ret.git.plugin.azure.PipelineRunState
-import io.rabobank.ret.git.plugin.azure.PullRequest
-import io.rabobank.ret.git.plugin.azure.Repository
-import io.rabobank.ret.git.plugin.azure.Reviewer
+import io.rabobank.ret.git.plugin.provider.azure.AzureDevopsClient
+import io.rabobank.ret.git.plugin.provider.azure.AzureResponse
+import io.rabobank.ret.git.plugin.provider.azure.Branch
+import io.rabobank.ret.git.plugin.provider.azure.Pipeline
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRun
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRunResult
+import io.rabobank.ret.git.plugin.provider.azure.PipelineRunState
+import io.rabobank.ret.git.plugin.provider.azure.PullRequest
+import io.rabobank.ret.git.plugin.provider.azure.Repository
+import io.rabobank.ret.git.plugin.provider.azure.Reviewer
 import io.rabobank.ret.git.plugin.command.AutoCompleteCommand
 import io.rabobank.ret.git.plugin.config.PluginConfig
 import io.rabobank.ret.git.plugin.output.OutputHandler
